moved classes to stack, and minor code cleanup
Former-commit-id: 33517fe3ad36dff13e6f91e6e319e124a7c6bd19

+++ engine/include/david/utils/utils.h
@@ -200,7 +200,7 @@ void printBoard(uint64_t board, const int index = -1);
 void yellDeprecated(const std::string info);
 
 bool perft();
-bool perft(const uint8_t limit, const uint8_t startDepth = 0);
+bool perft(const int limit, const uint8_t startDepth = 0);
 uint64_t perft(const uint8_t depth, const ::david::type::gameState_t& gs, std::array<int, 6>& results);
 const std::string getEGN(const ::david::type::gameState_t& first, const ::david::type::gameState_t& second);
 void getEGN(const ::david::type::gameState_t& first, const ::david::type::gameState_t& second, std::string& EGN);
